requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
mysql-connector==2.2.9
python-decouple==3.8
passlib==1.7.4
bcrypt==4.0.1
PyJWT==1.7.1
numpy==1.23.5
tensorflow==2.12.0
Pillow==9.5.0
python-multipart==0.0.6